fixing warning refs #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@14418 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSNet.cpp
@@ -359,7 +359,7 @@ MSNet::simulationStep() {
     // simulation state output
     std::vector<SUMOTime>::iterator timeIt = find(myStateDumpTimes.begin(), myStateDumpTimes.end(), myStep);
     if (timeIt != myStateDumpTimes.end()) {
-        const int dist = distance(myStateDumpTimes.begin(), timeIt);
+        const int dist = (int)distance(myStateDumpTimes.begin(), timeIt);
         MSStateHandler::saveState(myStateDumpFiles[dist], myStep);
     }
     myBeginOfTimestepEvents->execute(myStep);
